Refactor folder structure for better FastMCP integration
- Reorganized code into src/natural_shift_planner/ with modular structure
- Split monolithic files into focused modules:
  - api/ - FastAPI application layer (app, routes, schemas, etc.)
  - core/models/ - Domain models (employee, shift, schedule)
  - core/constraints/ - Optimization constraints
  - mcp/ - MCP server implementation
  - utils/ - Utility functions
- Updated all imports to work with new structure
- Fixed Python version to 3.12 for Timefold compatibility
- Moved tests to dedicated tests/ directory
- Updated documentation (README.md, CLAUDE.md)
- Removed unnecessary setup scripts and backup files

Closes #3
