--- SETUP_COMPLETE.sh
@@ -1,64 +1,0 @@
-#!/bin/bash
-
-# Grant execution permission
-chmod +x setup-docker-apple-silicon.sh
-
-echo "🍎 Shift Scheduler Dev Container Environment for Apple Silicon Mac is ready!"
-echo "================================================================================"
-echo ""
-echo "📁 Created Files:"
-echo "├── .devcontainer/          # Dev Container configuration"
-echo "├── .vscode/                # VS Code settings, debugging, and extensions"
-echo "├── api-test.http           # REST API tests"
-echo "├── setup-docker-apple-silicon.sh  # Apple Silicon optimization script"
-echo "└── Other project files"
-echo ""
-echo "🚀 How to Start Development Environment:"
-echo ""
-echo "【Method 1: VS Code Dev Container (Recommended)】"
-echo "1. Open project in VS Code:"
-echo "   code /projects/shared/shift-scheduler"
-echo ""
-echo "2. Open Command Palette (Cmd+Shift+P) and run:"
-echo "   'Dev Containers: Reopen in Container'"
-echo ""
-echo "3. After Dev Container starts, in terminal:"
-echo "   make setup  # Install dependencies"
-echo "   make run    # Start application"
-echo ""
-echo "【Method 2: Command Line】"
-echo "1. Setup Docker environment:"
-echo "   ./setup-docker-apple-silicon.sh"
-echo ""
-echo "2. Start Dev Container:"
-echo "   make dev-setup"
-echo ""
-echo "3. Open in VS Code:"
-echo "   code ."
-echo "   # Then run 'Dev Containers: Reopen in Container'"
-echo ""
-echo "🎯 Key Features:"
-echo "✅ Native Apple Silicon (M1/M2/M3) Support"
-echo "✅ Full Timefold Solver Integration"
-echo "✅ Complete VS Code Integration (Debug, Test, Format)"
-echo "✅ Weekly Work Hours Constraints, Skill Matching, Break Time Management"
-echo "✅ REST Client API Testing"
-echo "✅ Ultra-fast Package Management with uv"
-echo ""
-echo "🔗 Access URLs (After Dev Container starts):"
-echo "• FastAPI Server: http://localhost:8081"
-echo "• Swagger UI: http://localhost:8081/docs"
-echo "• ReDoc: http://localhost:8081/redoc"
-echo ""
-echo "📚 Documentation:"
-echo "• README.md - Comprehensive Guide"
-echo "• api-test.http - API Test Examples"
-echo ""
-echo "🛠 Common Commands (Inside Dev Container):"
-echo "make run      # Start server"
-echo "make test     # Run tests"
-echo "make format   # Format code"
-echo "make lint     # Check code"
-echo "make test-api # Verify API functionality"
-echo ""
-echo "🎉 Setup Complete! Enjoy your development experience!"
