--- finalize-setup.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Grant execution permission
-chmod +x SETUP_COMPLETE.sh
-
-# Display setup completion message
-./SETUP_COMPLETE.sh
