--- run-setup.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-chmod +x finalize-setup.sh
-./finalize-setup.sh
