+++ src/natural_shift_planner/api/__init__.py
@@ -1,0 +1,7 @@
+"""
+Shift Scheduler API
+"""
+from .app import app
+from .schemas import *  # noqa: F401, F403
+
+__all__ = ["app"]
