+++ src/natural_shift_planner/api/app.py
@@ -1,0 +1,21 @@
+"""
+FastAPI application instance and configuration
+"""
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+
+# Create FastAPI application
+app = FastAPI(
+    title="Shift Scheduler API",
+    description="Shift creation API using Timefold Solver",
+    version="1.0.0",
+)
+
+# CORS settings
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
