+++ src/natural_shift_planner/core/constraints/__init__.py
@@ -1,0 +1,6 @@
+"""
+Shift scheduling constraints
+"""
+from .shift_constraints import shift_scheduling_constraints
+
+__all__ = ["shift_scheduling_constraints"]
