+++ src/natural_shift_planner/core/constraints/shift_constraints.py
@@ -9,7 +9,7 @@ from timefold.solver.score import (
     constraint_provider,
 )
 
-from models import Employee, Shift, ShiftSchedule
+from ..models import Employee, Shift, ShiftSchedule
 
 
 @constraint_provider
