+++ src/natural_shift_planner/core/models/__init__.py
@@ -1,0 +1,8 @@
+"""
+Core domain models
+"""
+from .employee import Employee
+from .schedule import ShiftSchedule
+from .shift import Shift
+
+__all__ = ["Employee", "Shift", "ShiftSchedule"]
