+++ src/natural_shift_planner/mcp/__init__.py
@@ -1,0 +1,6 @@
+"""
+MCP server for Shift Scheduler
+"""
+from .server import mcp
+
+__all__ = ["mcp"]
