+++ src/natural_shift_planner/utils/__init__.py
@@ -1,0 +1,6 @@
+"""
+Utility functions
+"""
+from .demo_data import create_demo_schedule
+
+__all__ = ["create_demo_schedule"]
