--- uv.lock.backup
@@ -1,2 +1,0 @@
-# このファイルは削除されました
-# uv.lock は uv sync 実行時に自動生成されます
